+++ eSTOL_Training_Tool/eSTOL_Training_Tool_Core/UI/FormUI.cs
@@ -529,7 +529,7 @@ namespace eSTOL_Training_Tool_Core.UI
             this.buttonPauseUnpause.Visible = controller.isPaused;
             this.buttonUnFlip.Enabled = controller.plane.isInit && controller.plane.IsFlipped() && !controller.plane.isReadonly;
             this.buttonCreatePreset.Enabled = controller.plane.isInit && controller.plane.IsStopped();
-            this.buttonSetFuelStandard.Enabled = controller.plane.isInit && controller.plane.IsStopped();
+            this.buttonSetFuelStandard.Enabled = false; // controller.plane.isInit && controller.plane.IsStopped();
 
             // set pattern altitude
             if (controller.stol != null && controller.stol.preset != null)